requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
sqlmodel==0.0.8
python-dotenv==1.0.0
requests==2.31.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
razorpay==1.2.3
gunicorn==20.1.0
